Update TwoCoins.cpp
Reducing the number of iterations in the for loop while it was running actually caused the probability to be closer to 25% rather than 33%. This is because the amount of fair coins flipped and the amount of unfair coins flipped were still relatively equal, despite less repeats. (I can definitely see the problem with that after typing that sentence. (The probability for each of the three different outcomes are still the same.)) So fairHeads/flips was still closer to 1/4.* Now the amount of fairTails is subtracted from the amount of flips after the for loop is finished.

* There are three different outcomes, but it's not 1/3 because there's an uneven distribution of probability among them. (fairHeads < unfairHeads)